добавлена работа с Избранными событиями

+++ src/main/java/ru/technocracy/echoserver/repositories/NotedEventRepository.java
@@ -3,6 +3,7 @@ package ru.technocracy.echoserver.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.relational.core.sql.In;
 import org.springframework.data.repository.query.Param;
 import ru.technocracy.echoserver.models.notedevent.NoteType;
 import ru.technocracy.echoserver.models.notedevent.NotedEvent;
@@ -12,12 +13,12 @@ import java.util.List;
 public interface NotedEventRepository extends JpaRepository<NotedEvent, NotedEventPK> {
 
     @Query("SELECT nt.id.eventId FROM NotedEvent nt WHERE nt.id.userId =: userId")
-    public List<Long> getNotedEventsByUserId(@Param("userId")Long userId);
+    public List<Integer> getNotedEventsByUserId(@Param("userId")Long userId);
 
     @Query("SELECT nt.id.eventId FROM NotedEvent nt" +
             " WHERE nt.noteType =: status AND nt.id.userId =: userId")
-    public List<Long> getEventByUserIdWhereStatusIs(@Param("userId") Long userId,
-                                                         @Param("status") NoteType status);
+    public List<Integer> getEventByUserIdWhereStatusIs(@Param("userId") Long userId,
+                                                  @Param("status") NoteType status);
 
     public NotedEvent getNotedEventById(NotedEventPK notedEventPK);
